Simplified the test-group structure to make it always take an initializer thunk, which could possibly be the void function. This simplifies the model's use of the structure because it no longer has to check for #f.
Revised the procedure for recovering from a test calling (exit) or killing
its own thread. Now a custodian shuts down all resources that had been open.
This could possibly be a problem for tests down the line if they relied on
a prior test obtaining some resource, but in general I think this way is
better.