fixed a bug with the set!/define decision code... the earlier code used:
(define scope (make-object module-scope%))

(define (read-python path)
    (let ([ast-l (build-ast-from-file path)])
       (for-each (lambda (a) (send a set-bindings! scope)) ast-l)
       ast-l))

and this would not work for some reason... maybe because scope is a module-level variable?  It needs to be mutated inside set-bindings!, maybe that's why.

now it works this way:
(let ([ast-l (build-ast-from-file path)]
      [scope (make-object module-scope%)])
    (for-each (lambda (a) (send a set-bindings! scope)) ast-l)
    ast-l)


this is also more correct, because even if the first version had worked, the scope would not get a "clean start" for each call to read-python.