The install script is no complete, it can scan the "collects" directory, touching compiled .zo and native files, and modifies .dep files (only when the `+d' flag is used and no `-z'), and it can use a specified value for PLTHOME, so it is possible to run it to prepare for moving the whole directory elsewhere.  .dep files are changed after the bin scripts are changed, and the old PLTHOME value is used for this (.deps will change prefixes with that value to the new plthome).
In addition, "setup-go.ss" has an option to turn on
`trust-existing-zos' (from cm.ss) which is being used by default in
the install script.  This means that install will trust and .zo files
it finds, which should make it possible to distribute a PLT tree with
.zo files included, and if not then it will do the same thing as
before -- simply compiling them.